Python-Flask-API/requirements.txt
aniso8601==9.0.1
attrs==20.3.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
psycopg2==2.8.6
pyrsistent==0.17.3
python-dotenv==0.15.0
pytz==2021.1
six==1.15.0
SQLAlchemy==1.3.23
Werkzeug==0.16.0
